# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Matrix Builder API",
    "description": "Public orchestration API for controlled Matrix Bundle generation.",
    "version": "0.8.0-batch.8"
  },
  "paths": {
...
    "/api/v1/bundles": {
      "post": {
        "tags": [
          "bundles"
        ],
        "summary": "Generate Bundle",
        "operationId": "generate_bundle_api_v1_bundles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BundleGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatrixBundle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bundles/{bundle_id}": {
      "get": {
        "tags": [
          "bundles"
        ],
        "summary": "Get Bundle",
        "operationId": "get_bundle_api_v1_bundles__bundle_id__get",
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bundle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatrixBundle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bundles/{bundle_id}/manifest": {
      "get": {
        "tags": [
          "bundles"
        ],
        "summary": "Get Bundle Manifest",
        "operationId": "get_bundle_manifest_api_v1_bundles__bundle_id__manifest_get",
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bundle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundleManifest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bundles/{bundle_id}/tree": {
      "get": {
        "tags": [
          "bundles"
        ],
        "summary": "Get Bundle Tree",
        "operationId": "get_bundle_tree_api_v1_bundles__bundle_id__tree_get",
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bundle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BundleTreeNode"
                  },
                  "title": "Response Get Bundle Tree Api V1 Bundles  Bundle Id  Tree Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bundles/{bundle_id}/prompt/{coder}": {
      "get": {
        "tags": [
          "bundles"
        ],
        "summary": "Get Bundle Prompt",
        "operationId": "get_bundle_prompt_api_v1_bundles__bundle_id__prompt__coder__get",
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bundle Id"
            }
          },
          {
            "name": "coder",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Coder"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bundles/{bundle_id}/signed-url": {
      "post": {
        "tags": [
          "bundles"
        ],
        "summary": "Create Signed Bundle Url",
        "operationId": "create_signed_bundle_url_api_v1_bundles__bundle_id__signed_url_post",
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bundle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedBundleUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bundles/{bundle_id}/save": {
      "post": {
        "tags": [
          "bundles"
        ],
        "summary": "Save Bundle",
        "operationId": "save_bundle_api_v1_bundles__bundle_id__save_post",
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bundle Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BundleSaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundleSaveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bundles/{bundle_id}/validate": {
      "post": {
        "tags": [
          "bundles"
        ],
        "summary": "Validate Bundle",
        "operationId": "validate_bundle_api_v1_bundles__bundle_id__validate_post",
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bundle Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ValidationRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bundles/{bundle_id}/publish-to-matrixhub": {
      "post": {
        "tags": [
          "bundles"
        ],
        "summary": "Publish Bundle To Matrixhub",
        "operationId": "publish_bundle_to_matrixhub_api_v1_bundles__bundle_id__publish_to_matrixhub_post",
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bundle Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/PublicationRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bundles/{bundle_id}/download": {
      "get": {
        "tags": [
          "bundles"
        ],
        "summary": "Download Bundle",
        "operationId": "download_bundle_api_v1_bundles__bundle_id__download_get",
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bundle Id"
            }
          },
          {
            "name": "expires",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expires"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prompts/{coder}": {
      "get": {
        "tags": [
          "prompts"
        ],
        "summary": "Get Demo Prompt",
        "operationId": "get_demo_prompt_api_v1_prompts__coder__get",
        "parameters": [
          {
            "name": "coder",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Coder"
            }
          },
          {
            "name": "bundle_url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bundle Url"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prompts/{bundle_id}/{coder}": {
      "get": {
        "tags": [
          "prompts"
        ],
        "summary": "Get Prompt For Bundle",
        "operationId": "get_prompt_for_bundle_api_v1_prompts__bundle_id___coder__get",
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bundle Id"
            }
          },
          {
            "name": "coder",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Coder"
            }
          },
          {
            "name": "bundle_url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bundle Url"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prompts/{bundle_id}/pack/all": {
      "get": {
        "tags": [
          "prompts"
        ],
        "summary": "Get Prompt Pack",
        "operationId": "get_prompt_pack_api_v1_prompts__bundle_id__pack_all_get",
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bundle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptPack"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BundleGenerationRequest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "matrix.builder.bundle-generation/v1"
          },
          "idea_request": {
            "$ref": "#/components/schemas/IdeaRequest"
          },
          "candidate_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Id"
          },
          "preferred_coder": {
            "$ref": "#/components/schemas/CoderId",
            "default": "generic-ai-coder"
          },
          "persist": {
            "type": "boolean",
            "title": "Persist",
            "default": false
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "idea_request"
        ],
        "title": "BundleGenerationRequest"
      },
      "BundleManifest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "matrix.builder.bundle-manifest/v1"
          },
          "bundle_id": {
            "type": "string",
            "title": "Bundle Id"
          },
          "blueprint_id": {
            "type": "string",
            "title": "Blueprint Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "status": {
            "$ref": "#/components/schemas/BundleStatus",
            "default": "ready"
          },
          "manifest_digest": {
            "type": "string",
            "title": "Manifest Digest"
          },
          "zip_digest": {
            "type": "string",
            "title": "Zip Digest"
          },
          "zip_size_bytes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Zip Size Bytes"
          },
          "file_count": {
            "type": "integer",
            "minimum": 1.0,
            "title": "File Count"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/BundleFile"
            },
            "type": "array",
            "title": "Files"
          },
          "prompts_available": {
            "items": {
              "$ref": "#/components/schemas/CoderId"
            },
            "type": "array",
            "title": "Prompts Available"
          },
          "standards": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Standards"
          },
          "storage_uri": {
            "type": "string",
            "title": "Storage Uri"
          },
          "checksums": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Checksums"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "bundle_id",
          "blueprint_id",
          "title",
          "created_at",
          "manifest_digest",
          "zip_digest",
          "zip_size_bytes",
          "file_count",
          "files",
          "prompts_available",
          "storage_uri"
        ],
        "title": "BundleManifest"
      },
...
      "BundleStatus": {
        "type": "string",
        "enum": [
          "draft",
          "ready",
          "expired",
          "archived",
          "saved"
        ],
        "title": "BundleStatus"
      },
      "BundleTreeNode": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": true
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Digest"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "path",
          "kind"
        ],
        "title": "BundleTreeNode"
      },
...
      "CoderId": {
        "type": "string",
        "enum": [
          "claude-code",
          "codex-chatgpt",
          "cursor",
          "gitpilot",
          "ibm-bob",
          "generic-ai-coder"
        ],
        "title": "CoderId"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatrixBundle": {
        "properties": {
          "bundle_id": {
            "type": "string",
            "title": "Bundle Id"
          },
          "blueprint_id": {
            "type": "string",
            "title": "Blueprint Id",
            "default": "bp_demo_standard"
          },
          "status": {
            "$ref": "#/components/schemas/BundleStatus",
            "default": "ready"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "bundle_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bundle Url"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "signed_download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed Download Url"
          },
          "manifest_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest Url"
          },
          "manifest_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest Digest"
          },
          "zip_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zip Digest"
          },
          "zip_size_bytes": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Zip Size Bytes"
          },
          "file_count": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "File Count"
          },
          "expires_in_seconds": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expires In Seconds",
            "default": 172800
          },
          "persisted": {
            "type": "boolean",
            "title": "Persisted",
            "default": false
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "storage_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Uri"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/BundleFile"
            },
            "type": "array",
            "title": "Files"
          },
          "tree": {
            "items": {
              "$ref": "#/components/schemas/BundleTreeNode"
            },
            "type": "array",
            "title": "Tree"
          },
          "prompts_available": {
            "items": {
              "$ref": "#/components/schemas/CoderId"
            },
            "type": "array",
            "title": "Prompts Available"
          },
          "standards": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Standards"
          },
          "validation": {
            "$ref": "#/components/schemas/ValidationStatus",
            "default": "not-run"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Links"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "bundle_id",
          "title"
        ],
        "title": "MatrixBundle"
      },
...
      "SignedBundleUrlResponse": {
        "properties": {
          "bundle_id": {
            "type": "string",
            "title": "Bundle Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "expires_in_seconds": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expires In Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "bundle_id",
          "url",
          "expires_at",
          "expires_in_seconds"
        ],
        "title": "SignedBundleUrlResponse"
      },
...
      "ValidationStatus": {
        "type": "string",
        "enum": [
          "not-run",
          "approved",
          "needs-repair",
          "rejected"
        ],
        "title": "ValidationStatus"
      },
...
    }
  }
}